% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcew.R
\name{get_area}
\alias{get_area}
\title{Get QCEW Data by Area}
\usage{
get_area(year, qtr, area)
}
\arguments{
\item{year}{(character, or numeric) Possible values: "1990", "1991", ..., "2015".}

\item{qtr}{(character, or numeric) Possible values: "1", "2", "3", "4", and "a" for annual average.}

\item{area}{(character, or numeric) Possible values: \url{http://www.bls.gov/cew/doc/titles/area/area_titles.htm}}
}
\value{
Return a tibble, i.e., modern data frame.
}
\description{
A function takes a year, quarter, and area argument and
returns a data frame containing the associated area data.
}
\examples{
\dontrun{
michigan_data = qcewAPI::get_area("2013", "a", "26000")
michigan_data = qcewAPI::get_area(2013, "a", 26000)
}

}

