% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcew.R
\name{get_size}
\alias{get_size}
\title{Get QCEW Data by Establishment Size}
\usage{
get_size(year, size)
}
\arguments{
\item{year}{(character, or numeric) Possible values: "1990", "1991", ..., "2015".}

\item{size}{(character, or numeric) Possible values: \url{http://www.bls.gov/cew/doc/titles/size/size_titles.htm}}
}
\value{
Return a tibble, i.e., modern data frame.
}
\description{
A function takes a year and establishment size class code
and returns an array containing the associated size data. Size data
is only available for the first quarter of each year.
}
\examples{
\dontrun{
size_100_249 = qcewAPI::get_size("2013", "6")
size_100_249 = qcewAPI::get_size(2013, 6)
}

}

